Add a basic async client/server pair to handle saving worklogs
This commit adds a basic async client and server, alongside a first version of
different optional storage backends, in order to handle passing, saving, and
storing worklogs.

The server supports both UNIX and TCP sockets while its message handler
expects JSON messages to make sure we are not tied to any specific way
of generating and/or sending these messages while still providing a single
interface for handling received messages.

The first storage backend is a simple CSV writer that saves received worklogs
to a CSV file with limited facilities to control the messages' output
representation.