Add a rough first draft for a curses-based console frontend
This commit adds a (rather rough) first draft for a console frontend using
curses to efficiently choose both an action and a project that could then be
sent as a message to the server.

TODO: This still requires some refactoring, as well as a more permanent format
for projects metadata!