Add daemonized server run function, add python-daemon to Pipfile
This commit removes the ability to launch the server directly from the module
file, replacing it instead with a dedicated server run function that adds
proper UNIX daemon support as well as basic UNIX signal handling. These
changes are part of preparations to allow running the server as a system
service, e.g. via systemd.