Add support for PostgreSQL10+ IDENTITY column, amend worklog model
This commit changes the model for the "worklog" table to use an automatically
increasing integer as primary key, since the only thing that can logically be
unique is a combination of all other fields, so we might as well not set a
primary key for them at all. Furthermore, by using an auto-increasing primary
key, deletions become visible, since the primary key just counts up for the
next row regardless.