Fix config initialisation, fix subparser type, add test argument
This commit removes the old (and rather) hacky way of calling the object
directly, which meant we could not supply additional defaults during
initialisation. This commit also adds a "test" dummy argument, which can be
used for testing. This argument does not trigger any actual behaviour yet it
satisfies the requirement to supply an argument to the program.