Add option to enter project/ticket keys to frontend.py, amend models
This commit adds the option to attach a ticket number to a worklog via the
curses client. To make selecting tickets as quick and seamless as possible,
this commit also alters the project model to hold a project key which then
gets prefixed to the entered ticket number, e.g. to get Jira-like keys like
TEST-123. The worklog model has been changed accordingly to hold ticket keys.